update project card
