+++ src/components/cards/project-card.js
@@ -1,12 +1,24 @@
 import * as React from 'react';
-import AspectRatio from '@mui/joy/AspectRatio';
+
+// mui imports
 import Card from '@mui/joy/Card';
 import CardContent from '@mui/joy/CardContent';
+import CardCover from '@mui/joy/CardCover';
 import Typography from '@mui/joy/Typography';
+import Stack from '@mui/material/Stack';
 
 // component imports
 import { GithubButton } from '../buttons/social-media-icon-buttons';
-const ProjectCard = () => {
+import SkillPillGrid from '../skills/skill-pill-grid';
+
+const ProjectCard = ({
+  width,
+  name,
+  description,
+  skills,
+  githubLink,
+  imageUrl,
+}) => {
   return (
     <Card
       variant="outlined"
@@ -14,24 +26,34 @@ const ProjectCard = () => {
         backgroundColor: '#080E21',
         borderWidth: '1px',
         borderColor: '#001F54',
-        boxShadow: '0 8px 16px 0 rgba(0,0,0,0.4)',
         borderRadius: '10px',
+        boxShadow: '0 8px 16px 0 rgba(0,0,0,0.4)',
         color: 'white',
+        minHeight: '280px',
+        width: { width },
       }}
     >
-      <div>
+      <CardCover>
+        <img src={imageUrl} alt={description} />
+      </CardCover>
+      <CardCover
+        sx={{
+          background: 'rgba(0,0,0,0.8)',
+        }}
+      />
+      <CardContent>
         <Typography level="title-lg" textColor={'white'}>
-          Straight Path: The Game
+          {name}
         </Typography>
-      </div>
-      <AspectRatio minHeight="120px" maxHeight="200px">
-        <img src="/projects/straight-path.png" alt="" />
-      </AspectRatio>
-      <CardContent
-        orientation="horizontal"
-        sx={{ display: 'flex', flexDirection: 'row-reverse' }}
-      >
-        <GithubButton />
+        <Typography level={'p'} textColor="white">
+          {description}
+        </Typography>
+      </CardContent>
+      <CardContent sx={{ justifyContent: 'flex-end' }}>
+        <Stack direction={'row'}>
+          <SkillPillGrid skills={skills} />
+          <GithubButton link={githubLink} />
+        </Stack>
       </CardContent>
     </Card>
   );
